virtapi/requirements.txt
libvirt-python
xmltodict==0.10.1
pyyaml==3.12
requests==2.13.0
prettytable
lxml
tqdm
paramiko
